refactor(extract/mitre/cwe): move from extract/cwe/ to extract/mitre/cwe/
Aligns the CWE extractor location with the rest of the MITRE family
(``pkg/extract/mitre/cvrf``, ``pkg/extract/mitre/v4``,
``pkg/extract/mitre/v5``) and with the renamed wire identifiers
introduced in #828 (``SourceID = mitre-cwe``, CLI ``mitre-cwe``,
output ``extract/mitre/cwe/``).

The Go import path changes; ``pkg/extract/types/cwe/`` stays at the
top level since the extract-types tree has no ``mitre/`` namespace.

Goldens regenerated because ``DataSource.Raws`` records fixture
paths.
