+++ pkg/extract/mitre/cwe/cwe.go
