+++ pkg/extract/mitre/cwe/cwe_test.go
@@ -4,7 +4,7 @@ import (
 	"path/filepath"
 	"testing"
 
-	"github.com/MaineK00n/vuls-data-update/pkg/extract/cwe"
+	"github.com/MaineK00n/vuls-data-update/pkg/extract/mitre/cwe"
 	utiltest "github.com/MaineK00n/vuls-data-update/pkg/extract/util/test"
 )
