+++ pkg/extract/mitre/cwe/testdata/fixtures/category/355.json
