+++ pkg/extract/mitre/cwe/testdata/fixtures/view/699.json
