+++ pkg/extract/mitre/cwe/testdata/fixtures/weakness/1007.json
