+++ pkg/extract/mitre/cwe/testdata/golden/cwe/CWE-1007.json
