+++ pkg/extract/mitre/cwe/testdata/golden/cwe/CWE-355.json
