+++ pkg/extract/mitre/cwe/testdata/golden/cwe/CWE-699.json
