+++ pkg/extract/mitre/cwe/testdata/golden/datasource.json
